extract cell translate into custom hook
